// Ophalen van eventTypes
// Toevoegen van eventype link naar Events table

// Datum/Tijd object aangepast in Event = Date (niet

<button type="button" class="btn btn-primary btn-lg btn-block">
Block level button</button>

// Ophalen van objecten uit DB was met int, aangepast naar Long

// Artist en event workflow aanpassen naar UI concept

// Meenemen van artist ID over verschillende pagina's. IS ook nog nodig voor user ID

+++ Artmin/src/main/java/artmin/service/ClientService.java
@@ -25,7 +25,7 @@ public class ClientService {
     private ClientDao dao;
      
     // zoeken van client op basis van ID
-    public Client findById(int id) {
+    public Client findById(Long id) {
         return dao.findById(id);
     }
      
@@ -52,7 +52,7 @@ public class ClientService {
     }
  
     // verwijderen van gebruiker
-    public void deleteClientById(long id) {
+    public void deleteClientById(Long id) {
         dao.deleteClientById(id);
     }
